+++ rxandroidble/src/test/groovy/com/polidea/rxandroidble/MockLocationServicesStatus.groovy
@@ -2,12 +2,11 @@ package com.polidea.rxandroidble
 
 import com.polidea.rxandroidble.internal.util.LocationServicesStatus
 
-class MockLocationServicesStatus extends LocationServicesStatus {
+class MockLocationServicesStatus implements LocationServicesStatus {
     boolean isLocationPermissionOk = true
     boolean isLocationProviderOk = true
 
     MockLocationServicesStatus() {
-        super(null, null, 25, 25, false)
     }
 
     @Override