+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 1.5.0-SNAPSHOT
 * Added possibility to change default operation timeout (https://github.com/Polidea/RxAndroidBle/pull/321)
 * Fixed Dagger2 compatibility (https://github.com/Polidea/RxAndroidBle/pull/342 https://github.com/Polidea/RxAndroidBle/pull/348)
 * Fixed DisconnectionRouter leaking subscription to RxBleAdapterStateObservable (https://github.com/Polidea/RxAndroidBle/pull/353)
+* Improved Location Services status check (https://github.com/Polidea/RxAndroidBle/issues/327)
 
 Version 1.4.3
 * Log informing that the underlying semaphore in a QueueSemaphore has been interrupted will be printed only when the situation was unexpected.(https://github.com/Polidea/RxAndroidBle/issues/317)